# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MindSafe Mental Health Platform API
  version: 1.0.0
  description: |
    Privacy-first AI mental health platform. All endpoints under `/api` are served by the
    API Gateway (Express) which proxies ML operations to internal Python microservices.

    ## Authentication
    Most endpoints require a Bearer token in the `Authorization` header.
    Obtain tokens via `/api/login` or `/api/auth/google`. Access tokens expire in 15 minutes;
    use `/api/refresh-token` to get a new pair.

    ## Error Format
    All errors return `{ "message": "..." }` with an appropriate HTTP status code.
  contact:
    name: MindSafe Team
  license:
    name: Private
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://mindsafe-api.onrender.com
    description: Free-tier production (Render)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/api/login`. Expires in 15 minutes.
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: "Invalid or expired token"
...
    ChatbotRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          maxLength: 5000
          example: "I've been feeling really anxious about my exams"
        session_id:
          type: string
          nullable: true
          description: Existing session ID to continue a conversation
        style:
          type: string
          enum: [warm, balanced, concise]
          default: balanced
        use_name:
          type: boolean
          default: true
          description: Whether the bot addresses user by name
        use_memory:
          type: boolean
          default: true
          description: Whether the bot remembers conversation context
    ChatbotResponse:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        message_id:
          type: string
        response:
          type: string
          example: "I hear you, Priya. Exam anxiety is really common..."
        emotion_detected:
          type: string
          example: anxiety
        emotional_intensity:
          type: number
          format: float
          minimum: 0
          maximum: 1
        crisis_level:
          type: string
          enum: [low, medium, high, critical]
        crisis_resources:
          type: array
          items:
            type: object
        coping_strategies:
          type: array
          items:
            type: string
        requires_escalation:
          type: boolean
        timestamp:
          type: string
          format: date-time
    ChatMessageRequest:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, ai, assistant, system]
        content:
          type: string
          maxLength: 5000
        encryptedContent:
          type: object
          description: Client-encrypted envelope (alternative to plaintext content)
          properties:
            encrypted:
              type: boolean
            ciphertext:
              type: string

    # ── Mood ──
    MoodLogRequest:
      type: object
      required: [mood_score]
      properties:
        mood_score:
          type: integer
          minimum: 1
          maximum: 10
          example: 7
        mood_label:
          type: string
          example: happy
        notes:
          type: string
          nullable: true
          example: "Had a good therapy session today"
        emotion_scores:
          type: object
          nullable: true
          additionalProperties:
            type: number
          example: { "joy": 0.8, "sadness": 0.1 }
        logged_at:
          type: string
          format: date-time
          nullable: true
    MoodLogResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: string
        logged_date:
          type: string
          format: date
        mood_score:
          type: integer
        mood_label:
          type: string
        notes:
          type: string
        emotion_scores:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    WeeklyScoreResponse:
      type: object
      properties:
        user_id:
          type: string
        week_start:
          type: string
          format: date
        week_end:
          type: string
          format: date
        entries_count:
          type: integer
        average_mood:
          type: number
        consistency_ratio:
          type: number
        stability_score:
          type: number
        weekly_mental_health_score:
          type: number
...
    ProfileRequest:
      type: object
      required: [fullName, email, mobile]
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 100
          example: "Priya Sharma"
        email:
          type: string
          format: email
        mobile:
          type: string
          pattern: '^\+?\d{10,15}$'
        anonymousName:
          type: string
          maxLength: 80
        anonymousMode:
          type: boolean

    # ── Avatar ──
...
paths:
...
  /api/chatbot:
    post:
      tags: [Chatbot]
      summary: Send message to AI companion
      operationId: askChatbot
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatbotRequest"
            examples:
              basic:
                summary: Simple message
                value:
                  content: "I've been feeling really anxious about my exams"
              with_session:
                summary: Continue conversation
                value:
                  content: "What coping strategies do you recommend?"
                  session_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                  style: warm
      responses:
        "200":
          description: Chatbot response with emotion analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatbotResponse"
              example:
                session_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                response: "I hear you. Exam anxiety is really common..."
                emotion_detected: anxiety
                emotional_intensity: 0.72
                crisis_level: low
                coping_strategies:
                  - "Try the 4-7-8 breathing technique"
                  - "Break study sessions into 25-minute blocks"
                requires_escalation: false
        "400":
          description: content is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: Chatbot service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "504":
          description: Chatbot request timed out (20s)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/chat:
    post:
      tags: [Chatbot]
      summary: Save a chat message
      operationId: addChatMessage
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageRequest"
      responses:
        "200":
          description: Message saved
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: [Chatbot]
      summary: Get chat message history
      operationId: getChatMessages
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Array of chat messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        role:
                          type: string
                        content:
                          type: string
                        timestamp:
                          type: string
                          format: date-time

  # ══════════ MOOD TRACKING ══════════
  /api/moods/log:
    post:
      tags: [Mood Tracking]
      summary: Log today's mood
      operationId: logMood
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoodLogRequest"
            example:
              mood_score: 7
              mood_label: happy
              notes: "Had a good therapy session today"
              emotion_scores:
                joy: 0.8
                sadness: 0.1
      responses:
        "201":
          description: Mood logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodLogResponse"
        "400":
          description: mood_score must be between 1 and 10
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: Mood service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/moods/weekly-score:
    get:
      tags: [Mood Tracking]
      summary: Get weekly mental health score
      operationId: getWeeklyScore
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Weekly composite score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyScoreResponse"
...
  /api/profile:
    get:
      tags: [Profile]
      summary: Get user profile
      operationId: getProfile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile:
                    type: object
                    properties:
                      fullName:
                        type: string
                      email:
                        type: string
                      mobile:
                        type: string
                      anonymousName:
                        type: string
                      anonymousMode:
                        type: boolean
    post:
      tags: [Profile]
      summary: Create or update profile
      operationId: upsertProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileRequest"
      responses:
        "200":
          description: Profile saved
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Profile]
      summary: Delete account and all data
      operationId: deleteProfile
      security:
        - BearerAuth: []
      description: |
        Permanently deletes user account, profile, chat history, mood data, and all
        associated records. This action is irreversible (GDPR right to erasure).
      responses:
        "200":
          description: Account deleted

  # ══════════ AVATAR ══════════
...